# docs/api/openapi.yaml
{
  "openapi": "3.1.0",
  "info": {
    "title": "EduMind MVP 0.3 API contract",
    "version": "0.3.0",
    "description": "MVP 0.1/0.2已实现接口及MVP 0.3拟实现契约；新增路径仅为T004设计，运行端点由T012–T028交付。JSON是YAML 1.2子集。"
  },
  "servers": [{"url": "/"}],
...
  "paths": {
...
    "/api/learning-units/{learningUnitId}/classroom/debate": {"post": {"tags": ["Classroom"],"summary": "启动数组vs链表结构化演示","operationId": "startStructuredDebate","security": [{"sessionCookie": []}],"parameters": [{"$ref": "#/components/parameters/LearningUnitId"},{"$ref": "#/components/parameters/CsrfToken"},{"$ref": "#/components/parameters/IdempotencyKey"},{"$ref": "#/components/parameters/IfMatchClassroomRevision"}],"responses": {"401": {"$ref": "#/components/responses/Unauthorized"},"403": {"$ref": "#/components/responses/CsrfFailed"},"404": {"$ref": "#/components/responses/NotFound"},"409": {"description": "IDEMPOTENCY_CONFLICT 或 课堂revision不匹配；code=CLASSROOM_VERSION_CONFLICT。","content": {"application/json": {"schema": {"$ref": "#/components/schemas/Error"}}}},"422": {"$ref": "#/components/responses/ValidationError"},"503": {"$ref": "#/components/responses/ReviewUnavailable"},"200": {"description": "POST fetch+ReadableStream；结构化候选为临时数据，review_pass后debate_ready携带已发布结果ID；失败仅error/done。","content": {"text/event-stream": {"schema": {"type": "string"}}}}},"requestBody": {"required": true,"content": {"application/json": {"schema": {"$ref": "#/components/schemas/DebateRequest"}}}},"description": "一次结构化流式生成，三个视角及主持人总结整体审核；不并行生成。保存原课堂返回点。 流开始后Review不可用以error/done传达，不改HTTP状态；agent_start含operation_id，状态由GET课堂操作查询。"}},
    "/api/learning-units/{learningUnitId}/classroom/debate/{resultId}": {"get": {"tags": ["Classroom"],"summary": "读取已审核演示结果","operationId": "getStructuredDebate","security": [{"sessionCookie": []}],"parameters": [{"$ref": "#/components/parameters/LearningUnitId"},{"$ref": "#/components/parameters/DebateResultId"}],"responses": {"200": {"description": "成功。","content": {"application/json": {"schema": {"$ref": "#/components/schemas/DebateResult"}}}},"401": {"$ref": "#/components/responses/Unauthorized"},"404": {"$ref": "#/components/responses/NotFound"}}}},
    "/api/learning-units/{learningUnitId}/classroom/debate/{resultId}/exit": {"post": {"tags": ["Classroom"],"summary": "退出演示并恢复课堂","operationId": "exitStructuredDebate","security": [{"sessionCookie": []}],"parameters": [{"$ref": "#/components/parameters/LearningUnitId"},{"$ref": "#/components/parameters/DebateResultId"},{"$ref": "#/components/parameters/CsrfToken"},{"$ref": "#/components/parameters/IdempotencyKey"},{"$ref": "#/components/parameters/IfMatchClassroomRevision"}],"responses": {"200": {"description": "成功。","content": {"application/json": {"schema": {"$ref": "#/components/schemas/ClassroomSnapshot"}}}},"401": {"$ref": "#/components/responses/Unauthorized"},"403": {"$ref": "#/components/responses/CsrfFailed"},"404": {"$ref": "#/components/responses/NotFound"},"409": {"description": "IDEMPOTENCY_CONFLICT 或 课堂revision不匹配；code=CLASSROOM_VERSION_CONFLICT。","content": {"application/json": {"schema": {"$ref": "#/components/schemas/Error"}}}},"422": {"$ref": "#/components/responses/ValidationError"}},"description": "恢复scene_progress/模式/角色，message_cursor不回拨；若场景已更新则提示当前已审核版本。"}},
    "/api/learning-units/{learningUnitId}/classroom/debate/{resultId}/feedback": {"post": {"tags": ["Classroom"],"summary": "记录显式视角偏好","operationId": "recordPerspectiveFeedback","security": [{"sessionCookie": []}],"parameters": [{"$ref": "#/components/parameters/LearningUnitId"},{"$ref": "#/components/parameters/DebateResultId"},{"$ref": "#/components/parameters/CsrfToken"},{"$ref": "#/components/parameters/IdempotencyKey"}],"responses": {"202": {"description": "成功。","content": {"application/json": {"schema": {"$ref": "#/components/schemas/PreferenceReceipt"}}}},"401": {"$ref": "#/components/responses/Unauthorized"},"403": {"$ref": "#/components/responses/CsrfFailed"},"404": {"$ref": "#/components/responses/NotFound"},"409": {"$ref": "#/components/responses/IdempotencyConflict"},"422": {"$ref": "#/components/responses/ValidationError"}},"requestBody": {"required": true,"content": {"application/json": {"schema": {"$ref": "#/components/schemas/PerspectiveFeedback"}}}},"description": "仅本owner已发布结果；证据先提交、画像版本异步合并，不从浏览/停留推断，不直接改掌握度。"}},
...
  },
  "components": {
    "securitySchemes": {"sessionCookie": {"type": "apiKey", "in": "cookie", "name": "edumind_session", "description": "Secure、HttpOnly、SameSite=Lax、Host-only Cookie，2h 绝对有效期；数据库只存令牌哈希。"}},
    "parameters": {
      "CsrfToken": {"name": "X-CSRF-Token", "in": "header", "required": true, "schema": {"type": "string", "minLength": 64, "maxLength": 64}, "description": "写请求必须提供；服务端同时验证同源 Origin。"},
      "IdempotencyKey": {"name": "Idempotency-Key", "in": "header", "required": true, "schema": {"type": "string", "minLength": 16, "maxLength": 128}, "description": "客户端为一次创建操作生成的随机键；与认证身份和请求摘要绑定。"},
...
      "LearningUnitId": {"name": "learningUnitId", "in": "path", "required": true, "schema": {"$ref": "#/components/schemas/Identifier"}},
...
      "DebateResultId": {"name": "resultId","in": "path","required": true,"schema": {"$ref": "#/components/schemas/Identifier"}},
      "IfMatchClassroomRevision": {"name": "If-Match-Classroom-Revision","in": "header","required": true,"schema": {"type": "integer","minimum": 1},"description": "当前课堂revision。幂等命中先于版本检查。"},
...
    },
    "responses": {
...
      "Unauthorized": {"description": "会话缺失、失效或不属于当前用户。", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/Error"}}}},
      "CsrfFailed": {"description": "Origin 或 CSRF token 校验失败。", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/Error"}}}},
      "NotFound": {"description": "资源不存在，或不属于当前用户。", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/Error"}}}},
      "ValidationError": {"description": "请求未通过 schema 校验。", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/Error"}}}},
      "IdempotencyConflict": {"description": "同一幂等键被用于不同请求摘要。", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/Error"}}}},
...
      "ReviewUnavailable": {"description": "审核不可用；code=REVIEW_UNAVAILABLE，禁止正式发布。","content": {"application/json": {"schema": {"$ref": "#/components/schemas/Error"}}}}
    },
    "schemas": {
...
      "Identifier": {"type": "string", "pattern": "^[A-Za-z0-9_-]{8,64}$"},
...
      "KnowledgeNodeId": {"type": "string", "pattern": "^[a-z][a-z0-9-]{1,63}$"},
...
      "Error": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "code",
          "message",
          "retryable"
        ],
        "properties": {
          "code": {
            "type": "string",
            "enum": [
              "UNAUTHORIZED",
              "CSRF_FAILED",
              "NOT_FOUND",
              "VALIDATION_ERROR",
              "IDEMPOTENCY_CONFLICT",
              "PROFILE_VERSION_CONFLICT",
              "PATH_VERSION_CONFLICT",
              "GRAPH_NODE_NOT_FOUND",
              "QUIZ_SUBMISSION_INVALID",
              "PATH_UNREACHABLE",
              "ACCOUNT_CONFLICT",
              "RATE_LIMITED",
              "PROVIDER_UNAVAILABLE",
              "REVIEW_REJECTED",
              "OPERATION_CANCELED",
              "JOB_STATE_CONFLICT",
              "RENDER_UNAVAILABLE",
              "RENDER_FAILED",
              "RENDER_INTERRUPTED",
              "MEDIA_UNAVAILABLE",
              "MEDIA_NOT_READY",
              "CLASSROOM_VERSION_CONFLICT",
              "SCENE_VERSION_CONFLICT",
              "EVENT_CURSOR_INVALID",
              "EVENT_CURSOR_EXPIRED",
              "MESSAGE_CURSOR_INVALID",
              "MESSAGE_CURSOR_EXPIRED",
              "REVIEW_UNAVAILABLE",
              "CONTENT_RETRACTED",
              "ROLE_DISABLED",
              "DEBATE_UNAVAILABLE",
              "DEBATE_ALREADY_ACTIVE",
              "DEBATE_NOT_ACTIVE",
              "DEBATE_CONTEXT_UNAVAILABLE",
              "DEBATE_PUBLICATION_FAILED",
              "DEBATE_INVALID_OUTPUT",
              "DEBATE_INTERRUPTED"
            ]
          },
          "message": {
            "type": "string"
          },
          "retryable": {
            "type": "boolean"
          },
          "operation_id": {
            "$ref": "#/components/schemas/Identifier"
          }
        }
      },
...
      "ClassroomSnapshot": {"type":"object","additionalProperties":false,"properties":{"learning_unit_id":{"$ref":"#/components/schemas/Identifier"},"revision":{"type":"integer","minimum":1},"message_cursor":{"type":"integer","minimum":0},"scene_key":{"type":"string"},"scene_version":{"type":"integer","minimum":1},"scene_progress":{"type":"integer","minimum":0},"mode":{"type":"string","enum":["focus","interactive"]},"enabled_roles":{"type":"array","items":{"type":"string","enum":["beginner","advanced"]},"uniqueItems":true},"generation_id":{"$ref":"#/components/schemas/Identifier"},"paused":{"type":"boolean"},"detour":{"type":"object","additionalProperties":false,"properties":{"scene_key":{"type":"string"},"scene_version":{"type":"integer","minimum":1},"scene_progress":{"type":"integer","minimum":0},"kind":{"type":"string","enum":["prerequisite","debate"]},"target_node_id":{"$ref":"#/components/schemas/KnowledgeNodeId"},"path_version_id":{"$ref":"#/components/schemas/Identifier"},"mode":{"type":"string","enum":["focus","interactive"]},"enabled_roles":{"type":"array","items":{"type":"string","enum":["beginner","advanced"]}},"paused":{"type":"boolean"},"result_id":{"$ref":"#/components/schemas/Identifier"}},"required":["scene_key","scene_version","scene_progress"]},"version_changed":{"type":"boolean"}},"required":["learning_unit_id","revision","message_cursor","scene_key","scene_version","scene_progress","mode","enabled_roles","paused"]},
...
      "DebateRequest": {"type": "object","additionalProperties": false,"properties": {"preset": {"type": "string","const": "array-vs-linked-list"},"question": {"type": "string","minLength": 1,"maxLength": 1000}},"required": ["preset","question"]},
      "DebateResult": {"type":"object","additionalProperties":false,"properties":{"id":{"$ref":"#/components/schemas/Identifier"},"scene_key":{"type":"string"},"scene_version":{"type":"integer","minimum":1},"status":{"const":"published"},"perspectives":{"type":"object","additionalProperties":false,"properties":{"performance":{"type":"string"},"engineering":{"type":"string"},"academic":{"type":"string"}},"required":["performance","engineering","academic"]},"moderator_summary":{"type":"string"},"question":{"type":"string","minLength":1},"question_conditions":{"type":"object","additionalProperties":false,"required":["stated","unknown"],"properties":{"stated":{"type":"array","items":{"type":"string","minLength":1}},"unknown":{"type":"array","items":{"type":"string","minLength":1}}}},"graph_basis":{"type":"object","additionalProperties":false,"required":["graph_version","nodes","relation"],"properties":{"graph_version":{"type":"string"},"nodes":{"type":"array","minItems":2,"maxItems":2,"items":{"type":"object","additionalProperties":false,"required":["id","name","description","ai_context","learning_objectives"],"properties":{"id":{"type":"string","enum":["array","single-linked-list"]},"name":{"type":"string"},"description":{"type":"string"},"ai_context":{"type":"string"},"learning_objectives":{"type":"array","items":{"type":"string"}}}}},"relation":{"type":"object","additionalProperties":false,"required":["type","node_ids"],"properties":{"type":{"const":"SIMILAR_TO"},"node_ids":{"type":"array","minItems":2,"maxItems":2,"items":{"type":"string","enum":["array","single-linked-list"]}}}}}},"moderator":{"type":"object","additionalProperties":false,"required":["objective_conclusion","tradeoffs","learner_advice"],"properties":{"objective_conclusion":{"type":"string","minLength":1},"tradeoffs":{"type":"string","minLength":1},"learner_advice":{"type":"string","minLength":1}}},"candidate_schema_version":{"type":"string","minLength":1},"candidate_prompt_version":{"type":"string","minLength":1},"generation_model_id":{"type":"string","minLength":1},"review_version":{"type":"string","minLength":1},"review_model_id":{"type":"string","minLength":1},"correction_attempts":{"type":"integer","minimum":0}},"required":["id","scene_key","scene_version","status","question","question_conditions","graph_basis","perspectives","moderator","moderator_summary","candidate_schema_version","candidate_prompt_version","generation_model_id","review_version","review_model_id","correction_attempts"]},
      "PerspectiveFeedback": {"type": "object","additionalProperties": false,"properties": {"perspective": {"type": "string","enum": ["performance","engineering","academic"]},"feedback": {"type": "string","const": "helpful"}},"required": ["perspective","feedback"]},
      "PreferenceReceipt": {"type": "object","additionalProperties": false,"properties": {"evidence_id": {"$ref": "#/components/schemas/Identifier"},"profile_version": {"type": ["integer", "null"],"minimum": 1},"update_status": {"type": "string","enum": ["updated","pending","unchanged","failed"]}},"required": ["evidence_id","profile_version","update_status"]},
...
    }
  },
...
}